+++ src/domain/services/index.js
@@ -1,0 +1,2 @@
+export { default as LotteryMachine } from './LotteryMachine';
+export { default as lottoService } from './lottoService';
